GCP/requirements.txt
functions-framework
google-cloud-firestore
requests